// Front/src/app/actividade/model/actividade-tarefa.ts
export class Actividade_tarefa{
     id: number;
    nome: string;
     descricao: string;
     data_inicio: any;
     data_fim: any;
     base_anexo_id: number;
     actividade_id: number;
     actividade_tarefa_estado_id: number;
     actividade_tarefa_tipo_id: number;
     duracao: number;
     responsavel: number;
     user_id: number;
     estado: number;
  }